# schemas/openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: RaptorFlow API
  version: 0.1.0
servers:
  - url: https://api.raptorflow.example.com
paths:
...
  /api/v1/nudges:
    get:
      summary: List nudges
      responses:
        "200":
          description: Nudge list
    post:
      summary: Create nudge
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [user_id, nudge_type, priority, title, body, source_type, source_id]
              properties:
                user_id:
                  type: string
                nudge_type:
                  type: string
                priority:
                  type: string
                title:
                  type: string
                body:
                  type: string
                action_type:
                  type: string
                action_data:
                  type: object
                source_type:
                  type: string
                source_id:
                  type: string
      responses:
        "201":
          description: Nudge created
        "400":
          description: Invalid request
  /api/v1/nudges/{id}:
    get:
      summary: Get nudge by ID
      responses:
        "200":
          description: Nudge found
        "404":
          description: Nudge not found
...
  /api/v1/nudges/{id}/viewed:
    post:
      summary: Mark nudge as viewed
      responses:
        "200":
          description: Status updated
        "404":
          description: Nudge not found
  /api/v1/nudges/{id}/dismissed:
    post:
      summary: Dismiss nudge
      responses:
        "200":
          description: Status updated
        "404":
          description: Nudge not found
...